Add user context validation to command handlers

+++ ClotheoAPI.Application/Users/Commands/DeleteUser/DeleteUserCommandHandler.cs
@@ -1,3 +1,4 @@
+using ClotheoAPI.Application.Auth.Context;
 using ClotheoAPI.Domain.Entities;
 using ClotheoAPI.Domain.Exceptions;
 using ClotheoAPI.Domain.Repositories;
@@ -5,12 +6,21 @@ using MediatR;
 
 namespace ClotheoAPI.Application.Users.Commands.DeleteUser;
 
-public class DeleteUserCommandHandler(IUserRepository userRepository) : IRequestHandler<DeleteUserCommand>
+public class DeleteUserCommandHandler(IUserRepository userRepository, IUserContext userContext)
+    : IRequestHandler<DeleteUserCommand>
 {
     public async Task Handle(DeleteUserCommand request, CancellationToken cancellationToken)
     {
         var userToDelete = await userRepository.GetByIdAsync(request.Id)
             ?? throw new NotFoundException(nameof(User), request.Id);
+
+        var currentUser = userContext.GetCurrentUser()
+            ?? throw new UnauthorizedException();
+
+        if (currentUser.Id != userToDelete.Id && !currentUser.IsAdmin)
+        {
+            throw new ForbiddenException();
+        }
 
         await userRepository.DeleteAsync(userToDelete);
     }
